Add notification.onClicked handler with a named function
This replaces the anonymous inline function used as the handler for
`browser.notification.onClicked.addListener` with a named function, to
avoid re-registering the same handler on every call to `addHandlers`.

Note that `addEventListener` called repeatedly with the same arguments
will not register the same event listener multiple times (see the
`EventTarget.addEventListener` spec[0] for more info). This means that
using the same named function for each call to `addHandlers` will only
call that function once, where the previous code added a new anonymous
function for each call.

This fixes #234, where clicking a notification would result in multiple
tabs being opened, depending on the number of times that `addHandlers`
was called

[0]:
https://www.w3.org/TR/2000/REC-DOM-Level-2-Events-20001113/events.html#Events-EventTarget-addEventListener